feat(rest): add rest docs to resource server
- updated asciidocs
- added asciidocs plugin and maven resource plugin to pom
- changed test name classes
- added ConfigurationContext to test classes
