+++ rest/authorization-server/pom.xml
@@ -132,6 +132,7 @@
     </dependencyManagement>
 
     <build>
+        <finalName>resource</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
