+++ rest/resource-server/src/docs/asciidoc/api-guide.adoc
@@ -8,7 +8,7 @@
 //
 
 = sw360 REST API Guide
-Kai Tödter; Version {project-version}
+BETA; Version {project-version}
 :doctype: book
 :icons: font
 :source-highlighter: highlightjs
@@ -94,12 +94,12 @@ include::{snippets}/should_document_headers/response-headers.adoc[]
 Whenever an error response (status code >== 400) is returned, the body will contain a JSON object
 that describes the problem. The error object has the following structure:
 
-include::{snippets}/should_document_errors/response-fields.adoc[]
+include::{snippets}/should_document_error_internal_error/response-fields.adoc[]
 
 For example, a request that attempts to get a non-existent project will produce a
 `500 Internal Server Error` response:
 
-include::{snippets}/should_document_errors/http-response.adoc[]
+include::{snippets}/should_document_error_internal_error/http-response.adoc[]
 
 [[overview-hypermedia]]
 === Hypermedia
