+++ rest/resource-server/src/docs/asciidoc/index.adoc
@@ -1,0 +1,22 @@
+//
+// Copyright Siemens AG, 2017. Part of the SW360 Portal Project.
+//
+// All rights reserved. This configuration file is provided to you under the
+// terms and conditions of the Eclipse Distribution License v1.0 which
+// accompanies this distribution, and is available at
+// http://www.eclipse.org/org/documents/edl-v10.php
+//
+
+= SW360 REST API
+BETA; Version {project-version}
+:doctype: book
+:icons: font
+:source-highlighter: highlightjs
+
+SW360 is the upcoming tool for our OSS management and clearing activities.
+Here you find information about the REST API for SW360.
+The REST API is in early stage, link:https://github.com/sw360/sw360portal[feedback/issues] is very welcome.
+
+* link:api-guide.html[API Guide]
+* link:https://github.com/sw360/sw360portal[Git Repository]
+* link:https://localhost:8443/browser.html#/api[Interactive API Browser]
